// src/app/models/venue.ts
export class Venue {
    venue_uuid: string;
    user_uuid: string;
    venue_name: string;
    venue_location: string;
    venue_location_longitude: number;
    venue_location_latitude: number;
    venue_description: string;
    venue_contact_email: string;
    venue_contact_phone: number;
    venue_contact_website: string;
    venue_capacity_minimum: number;
    venue_capacity_maximum: number;
    venue_operating_hours: string;
    venue_restrictions	: string;
}